Added middleware support
Supports two types of middleware:

- Handler middleware, which wraps slog.Handlers, and can intercept each log Record
- ReplaceAttrs middleware, which can manipulate record attrs when the record is being handled by the delegate handler

ReplaceAttrs middleware needs to be configured when the Delegate Handler is being constructed, so Config (which has the primary responsibility for building delegate handlers) supports configuring a list of ReplaceAttr functions.  They will be called in the order than are added to the Config.

Handler middleware is installed into the Controller via the Use() and UseDefault() methods.

Also:
- Added package variables for the slog levels, plus 2 additional "all" and "off" levels
- Added AbbreviateLevel(), FormatTimes(), and SimpleTime() ReplaceAttrs