Snapshot support
Added a convenience method for creating a snapshot of the configuration of a handler, in the form of a function which, when called, will revert the state of the handler to that prior snapshot state.

In addition to refactoring this package around this function, its also useful in other tests which want to temporarily capture or modify logs for the duration of the test.

Also removed the locking around the buffer.  This write synchronization is part of the Handler contract.